string.h: ajout strlen
git-svn-id: https://www.etud.insa-toulouse.fr/svn/projet_tut_OS/trunk/tacos@172 ab7442c9-e926-469a-b309-6eee04cae01a

+++ include/string.h
@@ -3,7 +3,7 @@
 
 #include <types.h>
 
-void* memcpy(void* dest, const void* src, int size);
+void* memcpy(void* dest, const void* src, size_t size);
 
 int strcmp(const char *s1, const char *s2);
